Update concurrent log test
Share a log among several threads, and also share two AtomicUint
counters.

One counter acts as a crude method of syncronization: some threads will
successfully CAS on the counter, and others will not. This signals to
the test thread to abort or complete their log reservations.

Upon success, a thread that completes a reservation will increment the
success counter. This counter indicates the number of sucessful writes
to the shared log.

Finally, iterate through the log, and assert that the number of
successful entries exactly matches the success counter.

Refs #404